# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "教室排课、教室资源管理和冲突检测 RESTful API。",
        "title": "教室排课助手 API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/schedule-versions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule-versions"
                ],
                "summary": "List schedule versions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/schedule-versions/compare": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule-versions"
                ],
                "summary": "Compare two schedule versions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "source version id",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "target version id",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/schedule-versions/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule-versions"
                ],
                "summary": "Get one schedule version with its snapshot entries",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "version id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/schedule-versions/{id}/publish": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule-versions"
                ],
                "summary": "Publish a schedule version",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "version id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/schedule-versions/{id}/rollback": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule-versions"
                ],
                "summary": "Roll back to a schedule version, creating a new draft copy",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "source version id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}